+++ src/domain/access_token/access_token_test.go
@@ -21,6 +21,6 @@ func TestGetNewAccessToken(t *testing.T) {
 func TestAccessTokenIsExpired(t *testing.T) {
 	at := AccessToken{}
 	assert.True(t, at.IsExpired(), "empty access token should be expired by default")
-	at.Expires = time.Now().UTC().Add(3 * time.Hour).Unix()
+	at.Expires = int(time.Now().UTC().Add(3 * time.Hour).Unix())
 	assert.False(t, at.IsExpired(), "access token expiring three hours from now should not be expired")
 }
